Refactor max health variables
Change type from float to int for semantics

+++ include/game_engine/objects/character_object.h
@@ -12,7 +12,7 @@ class CharacterObject : public GameObject {
   void Deserialize(const nlohmann::json &json) override;
  protected:
   CharacterObject(PhysicsComponent *physics_component);
-  float max_health_ = 1;
+  int max_health_ = 1;
   float health_ = 0;
 };
