+++ include/ikaruga/core/objects/enemy/enemy_type.h
@@ -11,7 +11,7 @@
 namespace ikaruga::objects::enemy {
 class EnemyType : public game_engine::JsonSerializable {
  public:
-  EnemyType(float max_health,
+  EnemyType(int max_health,
             ikaruga::objects::enemy::movement::Pattern pattern,
             const std::vector<projectile::ProjectileType> &projectile_types,
             const ci::Color &color);
@@ -22,11 +22,11 @@ class EnemyType : public game_engine::JsonSerializable {
   void Serialize(nlohmann::json &json) const override;
   void Deserialize(const nlohmann::json &json) override;
 
-  float GetMaxHealth() const;
+  int GetMaxHealth() const;
   const std::vector<projectile::ProjectileType> &GetProjectileTypes() const;
   const ci::ColorT<float> &GetColor() const;
  private:
-  float max_health_;
+  int max_health_;
   movement::Pattern pattern_;
   std::vector<projectile::ProjectileType> projectile_types_;
   ci::ColorT<float> color_;
